feat(router): proxy for controller, builder, and registry
The router component is to be the point of entry for the entire
Deis cluster.

This commit makes the router:
* Forward requests to deis.domain on to the controller
* Forward connections on port 2222 to the builder
* Forward connections on port 5000 to the registry

This requires manually compiling nginx from source with the
tcp_proxy module and adding appropriate upstreams for builder
and registry.

Also, the default route was updated to proxy to the controller.

TESTING: rebuild the router
```console
$ make -C router build restart
```

And focus testing in two areas:
* That nothing was broken with the change to custom-built nginx (use Deis)
* That the router is properly proxying requests on pots 80, 2222, 5000

closes #535
replaces #974